parsing good

+++ include/fdf.h
@@ -68,6 +68,6 @@ t_point orthogonal_project_point(t_point point3D, float scale, t_point translate
 void render_grid(t_map *map, t_img *img, float scale, t_point translate);
 t_point get_point_at(t_map *map, int x, int y);
 void draw_line(t_img *img, t_point p0, t_point p1);
-int determine_color(float height);
+int determine_color(int height);
 
 #endif
